+++ mandu/src/components/organisms/Header.js
@@ -1,31 +1,8 @@
 import {List} from "../atoms/ListItems";
 import {TopNavListItem} from "../molecules/NavList";
-import {Link} from "react-router-dom";
-import {useDispatch, useSelector} from "react-redux";
-import cookie from "react-cookies";
-import {setId} from "../../redux/userSlice";
-import {useEffect} from "react";
+import UserNav from "../molecules/UserNav";
 
 const Header = () => {
-    const userId = useSelector((state) => state.user.id);
-    const dispatch = useDispatch();
-    const checkValidUser = () => {
-        const cookieAccessToken = cookie.load('access_token');
-        const cookieUserId = cookie.load('user_id');
-        if (cookieAccessToken && cookieUserId) {
-            dispatch(setId(cookieUserId));
-        }
-    }
-
-    const logout = () => {
-        cookie.remove('access_token', {path: '/'});
-        cookie.remove('user_id', {path: '/'});
-        dispatch(setId(null));
-    }
-
-    useEffect(() => {
-        checkValidUser();
-    }, []);
 
     return (
         <header className="mt-1 max-w-screen-xl flex items-center w-full">
@@ -38,27 +15,7 @@ const Header = () => {
                 <TopNavListItem link="/about">라이브</TopNavListItem>
                 <TopNavListItem link="/about">기획전</TopNavListItem>
             </List>
-            <div className="flex mb-1">
-                <div className="self-center">
-                    <Link to="/cart">
-                        <img alt="장바구니" className="h-7 mx-2" src="cart.png"/>
-                    </Link>
-                </div>
-                <div className="pl-7 border-l-2 ml-4 px-2 mt-0.5">
-                    {userId
-                        ?
-                        <div className="flex">
-                            <div className="my-2">{`유저 id : ${userId}`}</div>
-                            <button onClick={logout}>
-                                <img alt="로그아웃" className="h-4 mx-4" src="logout.png"/>
-                            </button>
-                        </div> :
-                        <Link to="/login"
-                              className="text-sm ">
-                            로그인
-                        </Link>}
-                </div>
-            </div>
+            <UserNav/>
         </header>
     );
 }
